New request/response style for async requests
Adds support for non fiber-aware requests for the new
request/response style that uses a single subscription.

Previous behavior can be enabled by passing `old_style_request: true`
when making the requests, but it is enabled by default.

Breaking change: the new request/response style does not return
the sid closure since it relies on a single subscription to handle
all the requests.

Signed-off-by: Waldemar Quevedo <wally@synadia.com>